requirements.txt
# Core dependencies
numpy>=1.21.0
box2d-py>=2.3.8
pygame>=2.0.0

# Deep Learning - PyTorch with CUDA support
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Web interface
flask>=2.0.0
flask-socketio>=5.0.0
websockets>=10.0

# Data science and visualization
matplotlib>=3.5.0
pandas>=1.3.0
scipy>=1.7.0

# ML experiment tracking
mlflow>=3.1.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter for MLflow

# System monitoring
psutil>=5.8.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.12.0 

Flask==3.0.3
numpy==1.26.4
pandas==2.2.2
pygame==2.5.2
pytest==8.2.0
box2d==2.3.10
